add open build service assertions
obs_project assertion checks if a project is created in the open build
service. Otherwise, it creates the project.

obs_request_accepted checks if there is a request to a certain project
for the current package that has been accepted. Otherwise, if there is
no request, it creates one.

succeded_packaging checks if a package has been build on a certain open
build service project, otherwise it fails.

submitted_rpm checks if a package has been submitted to a certain open
build service project (this already existed but has been modified).
Otherwise, it creates a new package and submits that.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>